Separate select_node and EmptyClusterError into common file, per code review
